pc4/team/round2-operation-jwt/challenge/server/webapp/requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer~=2.0.0
click==8.1.3
cryptography==37.0.3
ecdsa==0.17.0
exrex==0.10.5
Faker==13.14.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
openapi-schema-pydantic==1.2.3
orjson==3.7.3
packaging==21.3
peewee==3.15.0
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pydantic-factories==1.3.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.20.3
starlite==1.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
